server: get rid of use-before-def errors in entities by putting the repo within the Entity class;

+++ server/models/entities/access-fields.ts
@@ -15,6 +15,15 @@ import { assert, unwrap } from '../../helpers';
 
 @Entity('access_fields')
 export default class AccessField {
+    private static REPO: Repository<AccessField> | null = null;
+
+    private static repo(): Repository<AccessField> {
+        if (AccessField.REPO === null) {
+            AccessField.REPO = getRepository(AccessField);
+        }
+        return AccessField.REPO;
+    }
+
     @PrimaryGeneratedColumn()
     id!: number;
 
@@ -51,29 +60,29 @@ export default class AccessField {
             typeof accessId === 'number',
             'AccessField.create second arg should have "accessId" id property'
         );
-        const entity = repo().create(Object.assign({}, attributes, { userId }));
-        return await repo().save(entity);
+        const entity = AccessField.repo().create(Object.assign({}, attributes, { userId }));
+        return await AccessField.repo().save(entity);
     }
 
     static async find(userId: number, fieldId: number): Promise<AccessField | undefined> {
-        return await repo().findOne({ where: { id: fieldId, userId } });
+        return await AccessField.repo().findOne({ where: { id: fieldId, userId } });
     }
 
     static async all(userId: number): Promise<AccessField[]> {
-        return await repo().find({ userId });
+        return await AccessField.repo().find({ userId });
     }
 
     static async exists(userId: number, fieldId: number): Promise<boolean> {
-        const found = await repo().findOne({ where: { userId, id: fieldId } });
+        const found = await AccessField.repo().findOne({ where: { userId, id: fieldId } });
         return !!found;
     }
 
     static async destroy(userId: number, fieldId: number): Promise<void> {
-        await repo().delete({ userId, id: fieldId });
+        await AccessField.repo().delete({ userId, id: fieldId });
     }
 
     static async destroyAll(userId: number): Promise<void> {
-        await repo().delete({ userId });
+        await AccessField.repo().delete({ userId });
     }
 
     static async update(
@@ -81,16 +90,8 @@ export default class AccessField {
         fieldId: number,
         attributes: Partial<AccessField>
     ): Promise<AccessField> {
-        await repo().update({ userId, id: fieldId }, attributes);
+        await AccessField.repo().update({ userId, id: fieldId }, attributes);
         const updated = await AccessField.find(userId, fieldId);
         return unwrap(updated);
     }
-}
-
-let REPO: Repository<AccessField> | null = null;
-function repo(): Repository<AccessField> {
-    if (REPO === null) {
-        REPO = getRepository(AccessField);
-    }
-    return REPO;
 }
