+++ server/models/entities/accesses.ts
@@ -17,6 +17,15 @@ import { bankVendorByUuid } from '../../lib/bank-vendors';
 
 @Entity('access')
 export default class Access {
+    private static REPO: Repository<Access> | null = null;
+
+    private static repo(): Repository<Access> {
+        if (Access.REPO === null) {
+            Access.REPO = getRepository(Access);
+        }
+        return Access.REPO;
+    }
+
     @PrimaryGeneratedColumn()
     id!: number;
 
@@ -94,7 +103,7 @@ export default class Access {
 
     // Doesn't insert anything in db, only creates a new instance and normalizes its fields.
     static cast(args: Partial<Access>): Access {
-        return repo().create(args);
+        return Access.repo().create(args);
     }
 
     static async create(
@@ -102,33 +111,36 @@ export default class Access {
         { fields = [], ...other }: Partial<Access>
     ): Promise<Access> {
         const fieldsWithUserId: Partial<AccessField>[] = fields.map(field => ({
-            userId,
             ...field,
+            userId,
         }));
-        const entity = repo().create({ userId, ...other, fields: fieldsWithUserId });
-        const access = await repo().save(entity);
+        const entity = Access.repo().create({ userId, ...other, fields: fieldsWithUserId });
+        const access = await Access.repo().save(entity);
         return access;
     }
 
     static async find(userId: number, accessId: number): Promise<Access | undefined> {
-        return await repo().findOne({ where: { userId, id: accessId }, relations: ['fields'] });
+        return await Access.repo().findOne({
+            where: { userId, id: accessId },
+            relations: ['fields'],
+        });
     }
 
     static async all(userId: number): Promise<Access[]> {
-        return await repo().find({ where: { userId }, relations: ['fields'] });
+        return await Access.repo().find({ where: { userId }, relations: ['fields'] });
     }
 
     static async exists(userId: number, accessId: number): Promise<boolean> {
-        const found = await repo().findOne({ where: { userId, id: accessId } });
+        const found = await Access.repo().findOne({ where: { userId, id: accessId } });
         return !!found;
     }
 
     static async destroy(userId: number, accessId: number): Promise<void> {
-        await repo().delete({ userId, id: accessId });
+        await Access.repo().delete({ userId, id: accessId });
     }
 
     static async destroyAll(userId: number): Promise<void> {
-        await repo().delete({ userId });
+        await Access.repo().delete({ userId });
     }
 
     static async update(
@@ -139,7 +151,7 @@ export default class Access {
         if (typeof newAttributes.fields !== 'undefined') {
             throw new Error('API error: use AccessField model instead!');
         }
-        await repo().update({ userId, id: accessId }, newAttributes);
+        await Access.repo().update({ userId, id: accessId }, newAttributes);
         return unwrap(await Access.find(userId, accessId));
     }
 
@@ -147,14 +159,6 @@ export default class Access {
         userId: number,
         { uuid: vendorId }: { uuid: string }
     ): Promise<Access[]> {
-        return await repo().find({ where: { userId, vendorId }, relations: ['fields'] });
+        return await Access.repo().find({ where: { userId, vendorId }, relations: ['fields'] });
     }
-}
-
-let REPO: Repository<Access> | null = null;
-function repo(): Repository<Access> {
-    if (REPO === null) {
-        REPO = getRepository(Access);
-    }
-    return REPO;
 }
